Move NXSortModel to the widget collection
