+++ src/nexpy/gui/widgets.py
@@ -17,7 +17,7 @@ from matplotlib.patches import Circle, Ellipse, Polygon, Rectangle
 
 from .pyqt import QtCore, QtGui, QtWidgets
 from .utils import (report_error, boundaries, get_color, format_float,
-                    find_nearest)
+                    find_nearest, natural_sort)
 
 
 warnings.filterwarnings("ignore", category=cbook.mplDeprecation)
@@ -80,6 +80,17 @@ class NXStack(QtWidgets.QWidget):
         self.box.remove(label)
 
 
+class NXSortModel(QtCore.QSortFilterProxyModel):
+
+    def __init__(self, parent=None):
+        super(NXSortModel, self).__init__(parent)
+
+    def lessThan(self, left, right):
+        left_text = self.sourceModel().itemFromIndex(left).text()
+        right_text = self.sourceModel().itemFromIndex(right).text()
+        return natural_sort(left_text) < natural_sort(right_text)
+
+    
 class NXScrollArea(QtWidgets.QScrollArea):
     """Scroll area embedding a widget."""
